cel_packer.cpp

+++ src/cel_packer.cpp
@@ -448,7 +448,7 @@ pass7_api_to_bitstreams(const AbstractPackedImage &api_,
       {
         int word_offset;
 
-        word_offset = std::max(2,row.size_32bit());
+        word_offset = std::max((uint64_t)2,row.size_32bit());
         row.write(0,
                   api_.offset_width,
                   (word_offset - 2));
